[ggml] mul, add and sub ops implemented for tensors. i think it would be easier to try and implement a bert model in order to figure out which ops we should include in the binding. for instance, is view and concat needed?
